Finish up Docker fix
The docker-entrypoint script is fairly arcane, but invoking it with
exec and $@ seems like the most foolproof way to have it behave like
before. In addition, we have to now split up the creation of the worker
file (and permissions changes) and its symlinking (which must be done
as the owner of the database). Finally, since we override ENTRYPOINT,
we need to ensure to provide a similar default for CMD so the image
can be started without any arguments, but they can be provided if
so desired (to set PostgreSQL configuration).