Modernize docker-compose usage (#36)
This change modernizes our Docker usage. We now use version 2.1 of the
docker-compose file format. This isn't as modern as version 3, but
unless we intend to use Docker Swarm, there is no need to rely on that
version. In addition, version 2.1 allows us to specify that containers
which depend on each other are only brought up after the containers
they rely on are healthy, not just running. This simplifies our logic,
as we can guarantee PostgreSQL connections will be possible and Docker
events will not be missed.

Secondly, this removes the use of the Citus worker list file in favor
of master_(add|remove)_node. With this update, one of the last pieces
of code depending on this file is now gone.

Finally, the docker-compose have also been updated to respect the
variables COMPOSE_PROJECT_NAME and MASTER_EXTERNAL_PORT, permitting
running two docker-compose clusters at once.

Most of this logic is made possible by the new membership manager
image, which reimplements our membership management in Python using
the Docker SDK for that language.